requirements.txt
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.20.3
pandas>=1.2.0
scikit-learn>=0.24.0
tqdm>=4.60.0
Pillow>=8.0.0
tensorboard>=2.5.0
colorama>=0.4.0
timm==0.6.13
tifffile>=2021.0.0
h5py>=3.0.0